Move util classes to util packages.
