+++ checker/src/main/java/org/checkerframework/checker/formatter/FormatterTransfer.java
@@ -3,6 +3,7 @@ package org.checkerframework.checker.formatter;
 import javax.lang.model.element.AnnotationMirror;
 import org.checkerframework.checker.formatter.FormatterTreeUtil.Result;
 import org.checkerframework.checker.formatter.qual.ConversionCategory;
+import org.checkerframework.checker.formatter.util.FormatUtil;
 import org.checkerframework.dataflow.analysis.RegularTransferResult;
 import org.checkerframework.dataflow.analysis.TransferInput;
 import org.checkerframework.dataflow.analysis.TransferResult;
