+++ checker/src/test/java/org/checkerframework/checker/test/junit/I18nFormatterUnitTest.java
@@ -1,7 +1,7 @@
 package org.checkerframework.checker.test.junit;
 
-import org.checkerframework.checker.i18nformatter.I18nFormatUtil;
 import org.checkerframework.checker.i18nformatter.qual.I18nConversionCategory;
+import org.checkerframework.checker.i18nformatter.util.I18nFormatUtil;
 import org.junit.Assert;
 import org.junit.Test;
